+++ simplekit/common/__init__.py
@@ -1,1 +1,2 @@
 from .fs import sizeof_fmt
+from .math import percent